Reducing the severity of oper fec attribute get failure (#2924)
What I did
Reduced the severity of log when get attribute of oper fec failed.

Why I did it
The oper fec is not a mandatory attribute and may not be implemented by vendor SAI, in which case throwing an error is wrong. Reduced the severity of the log message to notice. This will also fix the checker issues in sonic-net/sonic-buildimage#16785