Fix memory leak and object copying bugs in orchagent (#3017)
* Update orchagent for Bookworm

There are two types of compiler errors. The first is GCC saying that in
a for-loop, if the variable used for the objects is const, then a
reference to the original object in the list can be used, instead of
making a copy. The second is that two variables could potentially be
used uninitialized. This appears to be a false positive to me, but I
assigned it to 0 regardless.