test_vnet.py refactor (#3352)
This PR refactors the test_vnet.py file. It separates the common functionality into a library file and keeps the tests in place.
What I did
Split up the test_vnet.py and put the common functionality into a library file.
Why I did it
The test_vnet.py has 29 tests. When these tests are run, it takes a log time for these tests to execute.
In addition, these tests run in the same VS container and some tests do not perform a full cleanup after the run. As a result, the next test have to deal with stale state.v The plan is to create a library so future tests can be added in a separate file