Use -fsanitize=address instead of -lasan for ASAN linker flags (#3344)
What I did
For the ASAN linker flags, use -fsanitize=address instead of manually specifying -lasan.

Why I did it
This allows the compiler to use some other linkages instead of -lasan in case that's not the correct one for the platform. It (should) also make sure that the ASAN library is the first library that gets linked and loaded at runtime.
This should also fix a flaky issue with the p4orch tests failing to run because ASAN isn't loaded first (not sure why this is flaky though, it should be pretty consistent).