[vlanmgrd]: Preventing VLAN interfaces from going down when all of their member ports are down (#3370)
What I did
Started up the "dummy" interface that is already connected to the bridge named "Bridge" so that when all ports connected to the bridge are DOWN, the bridge interface remains UP. This ensures that when there is one VLAN interface connected to the bridge, it remains UP even when all of the VLAN's member ports are DOWN.
Since the VLAN interface is UP immediately after creation (because the "dummy" interface is UP), at least one route to the VLAN's subnet is added to the ASIC_STATE:SAI_OBJECT_TYPE_ROUTE_ENTRY table (one route if the VLAN is created in a peerless Vnet, and two routes if it is created in a VNet with one peer). Tests in tests/test_vnet.py have been updated to reflect this fact.

Why I did it
Before this change, when all member ports of a VLAN interface went down, the VLAN interface also went operationally down. This effectively disabled the advertisement of the VLAN interface's subnet IP to neighbors and prevented the IP-in-IP decapsulation feature for the VLAN interface from working properly.