[portsorch] query SAI_HOSTIF_ATTR_QUEUE switch capability once (#3260)
What I did

Support for SAI_HOSTIF_ATTR_QUEUE is a switch level capability, no need to query it per port.

Why I did it

As part of the goal to optimize boot time for systems with many ports. This one is the first step into that direction, extracting all non per-port operations.